requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
pandas>=2.0.0
python-dateutil>=2.8.0
requests>=2.25.0